Break out unit test & get them to pass
